pyproject.toml
[project]
name = "automaton_auditor"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "pydantic",
    "pypdf",
    "python-dotenv",
    "docling",
    "tree-sitter",
    "tree-sitter-python",
    "docling-core>=2.66.0",
    "langchain-ollama>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
